# backend/internal/api/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: |-
    Platform API to order and manage cloud resources — managed databases, apps, VMs (app servers), AI models, and public API endpoints — on the DADA Cloud platform.
    Mutations are asynchronous: create/update/delete enqueue an operation and return 202 Accepted with an operation object; poll GET /projects/{projectId}/operations/{operationId} until the operation reaches a terminal status (Ready or Failed).
  title: DADA Cloud Console API
  version: "1.0"
paths:
  /admin/audit:
    get:
      description: Returns audit_events rows joined with the actor's email and the
        project's display name, newest first. Filters by exact action and a case-insensitive
        email substring. Platform-admin only (/platform-admins group); every other
        caller gets 403.
      operationId: listAuditEvents
      parameters:
      - description: Exact audit_events.action value
        in: query
        name: action
        type: string
      - description: Case-insensitive substring match on the actor's email
        in: query
        name: user
        type: string
      - description: Max rows to return (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: object with an events array and a total count
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List audit events (platform-admin only)
      tags:
      - admin
  /admin/operations:
    get:
      description: Returns operations in the WaitingForApproval state across all projects
        where the caller is a platform-admin, each with the project name and requester.
        Read-only. The GPU-model creation gate is the first consumer.
      operationId: listAdminApprovals
      produces:
      - application/json
      responses:
        "200":
          description: object with an approvals array
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List operations awaiting admin approval
      tags:
      - admin
  /admin/operations/{opId}/approve:
    post:
      consumes:
      - application/json
      description: Approves a WaitingForApproval operation, transitioning it to Created
        so the dispatcher executes it. Platform-admin only. An optional note is recorded
        in the audit trail.
      operationId: approveOperation
      parameters:
      - description: Operation UUID
        in: path
        name: opId
        required: true
        type: string
      - description: Optional object with a note string
        in: body
        name: body
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: object with the updated operation
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve an operation awaiting approval
      tags:
      - admin
  /admin/operations/{opId}/reject:
    post:
      consumes:
      - application/json
      description: Rejects a WaitingForApproval operation, transitioning it to Cancelled.
        Platform-admin only. A reason is required and surfaced to the requester in
        the operations timeline.
      operationId: rejectOperation
      parameters:
      - description: Operation UUID
        in: path
        name: opId
        required: true
        type: string
      - description: Object with a required reason string
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: object with the updated operation
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject an operation awaiting approval
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"